Rerun analysis with "complete" gene set, including the one signaling gene
In the previous 2/8 I had excluded a single Signaling gene with outlier
dN/dS. But this was an "outlier removal" run to confirm it did make things
significant. However, it was not the results set intended to be presented
in the paper. So I needed to rerun including that one Signaling gene.