Correcting renaming of html result files
I previously used "git mv" to rename the html output file, in an
effort to preserve previous versions of the file. But then this
ended up getting overwritten by newer versions. So I checked out
the ealier version, correctly reflecting 2/8 time stamp, duplicated
it, and also duplicated the 3/3 version with specific names so
they are "static". Then I "git mv" the one file that is the expected
output from running the markdown file back to the original name.